requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
aiosqlite==0.17.0
anyio==3.3.2
asgiref==3.4.1
async-timeout==3.0.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.2
dnspython==2.1.0
email-validator==1.1.3
fakeredis==1.6.1
fastapi==0.70.0
fastapi-mail==1.0.1
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.2
iso8601==0.1.16
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
passlib==1.7.4
psycopg2-binary==2.9.1
pycparser==2.20
pydantic==1.8.2
PyJWT==2.2.0
pyparsing==2.4.7
pypika-tortoise==0.1.1
python-dotenv==0.19.1
python-multipart==0.0.5
pytz==2021.3
redis==3.5.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
starlette==0.16.0
tortoise-orm==0.17.8
typing-extensions==3.10.0.2
uvicorn==0.15.0
